requirements.txt
google-generativeai>=0.7.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0
scikit-learn>=1.3.0